% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexdiff.R
\name{git_latexdiff}
\alias{git_latexdiff}
\title{Call latexdiff on git revisions}
\usage{
git_latexdiff(path, revision, clean = TRUE, ...)
}
\arguments{
\item{path}{File path to diff}

\item{revision}{Revision, specified in a form that \code{git} understands. See
\code{man gitrevisions}}

\item{clean}{Clean up intermediate files, including the checked out file?}

\item{...}{Arguments passed to \code{\link[=latexdiff]{latexdiff()}}}
}
\value{
The result of \code{latexdiff}
}
\description{
\code{git_latexdiff()} checks out a previous version of a file and calls latexdiff
on it.
}
\examples{
\dontrun{
git_latexdiff("file1.Rmd", "HEAD^")
}
}
