% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexdiff.R
\name{latexdiff}
\alias{latexdiff}
\title{Produce a diff of two Rmd files using latexdiff}
\usage{
latexdiff(path1, path2, output = "diff", open = interactive())
}
\arguments{
\item{path1}{Path to the first file}

\item{path2}{Path to the second file}

\item{output}{Path to the output, without the ".tex" file extension.}

\item{open}{Logical. Automatically open the resulting PDF?}
}
\value{
Invisible NULL.
}
\description{
Produce a diff of two Rmd files using latexdiff
}
\examples{
\dontrun{
latexdiff("file1.Rmd", "file2.Rmd")
}
}
